requirements.txt
# agents==1.4.0
# carla==0.9.15
networkx==2.6.3
numpy==1.21.5
pygame==2.6.0
python-dotenv==0.21.1
# torch==1.13.0 
torchrl==0.1.1
